+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Instructions
   * Fix spelling mistake [#243](https://github.com/demokratie-live/democracy-client/issues/243) [#235](https://github.com/demokratie-live/democracy-client/issues/235)
+  * Swipe on click blue circle [#237](https://github.com/demokratie-live/democracy-client/issues/237)
 
 ### 0.6.5